rename some things and test for edge cases with clients
