+++ test/dump/clientConnectionsSpec.js
@@ -1,5 +1,7 @@
 import R from 'ramda'
-import { associate, disassociate } from '../../src/utils/associateClientConnections'
+import disassociate from '../../src/utils/clientConnections/disassociate'
+import associate from '../../src/utils/clientConnections/associate'
+
 import connections from '../fixtures/connections.json'
 import clients from '../fixtures/clients.json'
 
@@ -19,6 +21,15 @@ describe("connection to client mapping", () => {
     it('can deal with multiple clients', () => {
       expect(second.enabled_clients).to.eql(["google-auth", "machines-rule"]);
     });
+
+    it('ignores when missing', () => {
+      const modified = R.dissocPath([0, 'enabled_clients'], connections);
+      const fixture = { connections: modified, clients };
+
+      const { connections: assocConnections } = associate(fixture);
+
+      expect(assocConnections[0].enabled_clients).to.be.undefined;
+    });
   });
 
   context('disassociate', () => {
@@ -31,6 +42,17 @@ describe("connection to client mapping", () => {
 
     it('can deal with multiple clients', () => {
       expect(second.enabled_clients).to.eql(origSecond.enabled_clients);
+    });
+
+    it('ignores when missing', () => {
+      const modified = R.compose(
+        R.dissocPath(['manifest', 'connections', 0, 'enabled_clients']),
+        R.dissocPath(['connections', 0, 'enabled_clients'])
+      )(withManifest);
+
+      const result = disassociate(modified);
+
+      expect(result.manifest.connections[0].enabled_clients).to.be.undefined;
     });
   });
 });
